Correctly identify whizbang hero
Apparently, the Power log contains a random hero first, which is then later switched out for the selected whizbang hero.

In principle, the correct way of identifying the hero seems to be to scan for the hero enity in the create game section, but given that we want to use the whizbang code for other things as well, this decodes the hero from the whizbang deck code.