Identify and import the players whizbang deck
This uses a hardcoded list of deck identifiers (which will have to be updated for each new expansion).

It might be possible to achieve the same thing by consuming the logs during initialization, but I assumed that it would be complicated, and did not investigate further. This approach only required copying all the deck strings, and playing whizbang once per deck to figure out the mappings.