Review of whizbang-support.
When looking for whizbang deckcode in power.log we need to identify first
which PlayerID entity does tag=WHIZBANG_DECK_ID belongs to.

Using powerState as identifier doesn't help as it will always be on
heroType1State when finding tag=WHIZBANG_DECK_ID. It will fail when the
player use whizbang as playerID 2 or when the enemy use whizbang as
playerID 2.